# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Challenge Corrézien Disc-Golf API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /years:
    get:
      operationId: listYears
      tags: [years]
      summary: List all competition years
      responses:
        "200":
          description: List of years
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Year"
    post:
      operationId: createYear
      tags: [years]
      summary: Create a competition year
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateYearBody"
      responses:
        "201":
          description: Year created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Year"
  /years/{yearId}:
    delete:
      operationId: deleteYear
      tags: [years]
      summary: Delete a competition year
      parameters:
        - name: yearId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Year deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /stages:
    get:
      operationId: listStages
      tags: [stages]
      summary: List stages
      parameters:
        - name: yearId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stage"
    post:
      operationId: createStage
      tags: [stages]
      summary: Create a stage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStageBody"
      responses:
        "201":
          description: Stage created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stage"
  /stages/{stageId}:
    put:
      operationId: updateStage
      tags: [stages]
      summary: Update a stage
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStageBody"
      responses:
        "200":
          description: Stage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stage"
    delete:
      operationId: deleteStage
      tags: [stages]
      summary: Delete a stage
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stage deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /scoring-rules:
    get:
      operationId: listScoringRules
      tags: [scoring-rules]
      summary: List scoring rules
      responses:
        "200":
          description: List of scoring rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoringRule"
    post:
      operationId: createScoringRule
      tags: [scoring-rules]
      summary: Create a scoring rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoringRuleBody"
      responses:
        "201":
          description: Scoring rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringRule"
  /scoring-rules/{ruleId}:
    put:
      operationId: updateScoringRule
      tags: [scoring-rules]
      summary: Update a scoring rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoringRuleBody"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringRule"
    delete:
      operationId: deleteScoringRule
      tags: [scoring-rules]
      summary: Delete a scoring rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rule deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Year:
      type: object
      properties:
        id:
          type: integer
        year:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - year
        - createdAt
    CreateYearBody:
      type: object
      properties:
        year:
          type: integer
      required:
        - year
...
    Stage:
      type: object
      properties:
        id:
          type: integer
        yearId:
          type: integer
        name:
          type: string
        location:
          type: string
        date:
          type: string
        par:
          type: integer
        order:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - yearId
        - name
        - par
        - order
        - createdAt
    CreateStageBody:
      type: object
      properties:
        yearId:
          type: integer
        name:
          type: string
        location:
          type: string
        date:
          type: string
        par:
          type: integer
        order:
          type: integer
      required:
        - yearId
        - name
        - par
        - order
...
    ScoringRule:
      type: object
      properties:
        id:
          type: integer
        position:
          type: integer
        points:
          type: integer
      required:
        - id
        - position
        - points
    CreateScoringRuleBody:
      type: object
      properties:
        position:
          type: integer
        points:
          type: integer
      required:
        - position
        - points
...