# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Challenge Corrézien Disc-Golf API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Admin logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current admin session
 */
export const GetMeResponse = zod.object({
  isAuthenticated: zod.boolean(),
  username: zod.string().optional(),
});

/**
 * @summary List all competition years
 */
export const ListYearsResponseItem = zod.object({
  id: zod.number(),
  year: zod.number(),
  createdAt: zod.string(),
});
export const ListYearsResponse = zod.array(ListYearsResponseItem);

/**
 * @summary Create a competition year
 */
export const CreateYearBody = zod.object({
  year: zod.number(),
});

/**
 * @summary Delete a competition year
 */
export const DeleteYearParams = zod.object({
  yearId: zod.coerce.number(),
});

export const DeleteYearResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all players
 */
export const ListPlayersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  createdAt: zod.string(),
});
export const ListPlayersResponse = zod.array(ListPlayersResponseItem);

/**
 * @summary Create a player
 */
export const CreatePlayerBody = zod.object({
  name: zod.string(),
});

/**
 * @summary Update a player
 */
export const UpdatePlayerParams = zod.object({
  playerId: zod.coerce.number(),
});

export const UpdatePlayerBody = zod.object({
  name: zod.string(),
});

export const UpdatePlayerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a player
 */
export const DeletePlayerParams = zod.object({
  playerId: zod.coerce.number(),
});

export const DeletePlayerResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List stages
 */
export const ListStagesQueryParams = zod.object({
  yearId: zod.coerce.number().optional(),
});

export const ListStagesResponseItem = zod.object({
  id: zod.number(),
  yearId: zod.number(),
  name: zod.string(),
  location: zod.string().optional(),
  date: zod.string().optional(),
  par: zod.number(),
  order: zod.number(),
  createdAt: zod.string(),
});
export const ListStagesResponse = zod.array(ListStagesResponseItem);

/**
 * @summary Create a stage
 */
export const CreateStageBody = zod.object({
  yearId: zod.number(),
  name: zod.string(),
  location: zod.string().optional(),
  date: zod.string().optional(),
  par: zod.number(),
  order: zod.number(),
});

/**
 * @summary Update a stage
 */
export const UpdateStageParams = zod.object({
  stageId: zod.coerce.number(),
});

export const UpdateStageBody = zod.object({
  yearId: zod.number(),
  name: zod.string(),
  location: zod.string().optional(),
  date: zod.string().optional(),
  par: zod.number(),
  order: zod.number(),
});

export const UpdateStageResponse = zod.object({
  id: zod.number(),
  yearId: zod.number(),
  name: zod.string(),
  location: zod.string().optional(),
  date: zod.string().optional(),
  par: zod.number(),
  order: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a stage
 */
export const DeleteStageParams = zod.object({
  stageId: zod.coerce.number(),
});

export const DeleteStageResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List scores
 */
export const ListScoresQueryParams = zod.object({
  stageId: zod.coerce.number().optional(),
  playerId: zod.coerce.number().optional(),
});

export const ListScoresResponseItem = zod.object({
  id: zod.number(),
  playerId: zod.number(),
  stageId: zod.number(),
  score: zod.number(),
  points: zod.number(),
  playerName: zod.string().optional(),
  stageName: zod.string().optional(),
  createdAt: zod.string(),
});
export const ListScoresResponse = zod.array(ListScoresResponseItem);

/**
 * @summary Create or update a score
 */
export const CreateOrUpdateScoreBody = zod.object({
  playerId: zod.number(),
  stageId: zod.number(),
  score: zod.number(),
});

export const CreateOrUpdateScoreResponse = zod.object({
  id: zod.number(),
  playerId: zod.number(),
  stageId: zod.number(),
  score: zod.number(),
  points: zod.number(),
  playerName: zod.string().optional(),
  stageName: zod.string().optional(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a score
 */
export const DeleteScoreParams = zod.object({
  scoreId: zod.coerce.number(),
});

export const DeleteScoreResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get rankings for a year (stage breakdown + cumulated)
 */
export const GetRankingsParams = zod.object({
  yearId: zod.coerce.number(),
});

export const GetRankingsResponse = zod.object({
  yearId: zod.number(),
  year: zod.number(),
  stages: zod.array(
    zod.object({
      id: zod.number(),
      yearId: zod.number(),
      name: zod.string(),
      location: zod.string().optional(),
      date: zod.string().optional(),
      par: zod.number(),
      order: zod.number(),
      createdAt: zod.string(),
    }),
  ),
  stageRankings: zod.record(
    zod.string(),
    zod.array(
      zod.object({
        position: zod.number(),
        playerId: zod.number(),
        playerName: zod.string(),
        score: zod.number(),
        diffToPar: zod.number(),
        points: zod.number(),
      }),
    ),
  ),
  playerRankings: zod.array(
    zod.object({
      position: zod.number(),
      playerId: zod.number(),
      playerName: zod.string(),
      stageScores: zod.array(
        zod.object({
          stageId: zod.number(),
          stageName: zod.string(),
          score: zod.number(),
          par: zod.number(),
          diffToPar: zod.number(),
          points: zod.number(),
        }),
      ),
      totalScore: zod.number(),
      totalPar: zod.number(),
      totalDiffToPar: zod.number(),
      totalPoints: zod.number(),
    }),
  ),
  cumulatedRankings: zod.array(
    zod.object({
      position: zod.number(),
      playerId: zod.number(),
      playerName: zod.string(),
      totalScore: zod.number(),
      totalPar: zod.number(),
      totalDiffToPar: zod.number(),
      totalPoints: zod.number(),
    }),
  ),
});

/**
 * @summary Export rankings as CSV
 */
export const ExportRankingsCsvParams = zod.object({
  yearId: zod.coerce.number(),
});

/**
 * @summary List scoring rules
 */
export const ListScoringRulesResponseItem = zod.object({
  id: zod.number(),
  position: zod.number(),
  points: zod.number(),
});
export const ListScoringRulesResponse = zod.array(ListScoringRulesResponseItem);

/**
 * @summary Create a scoring rule
 */
export const CreateScoringRuleBody = zod.object({
  position: zod.number(),
  points: zod.number(),
});

/**
 * @summary Update a scoring rule
 */
export const UpdateScoringRuleParams = zod.object({
  ruleId: zod.coerce.number(),
});

export const UpdateScoringRuleBody = zod.object({
  position: zod.number(),
  points: zod.number(),
});

export const UpdateScoringRuleResponse = zod.object({
  id: zod.number(),
  position: zod.number(),
  points: zod.number(),
});

/**
 * @summary Delete a scoring rule
 */
export const DeleteScoringRuleParams = zod.object({
  ruleId: zod.coerce.number(),
});

export const DeleteScoringRuleResponse = zod.object({
  message: zod.string(),
});